数据采集接口调试

+++ src/main/java/com/fs/smartTown/modules/dataRegister/entity/EpidemicSurveillance.java
@@ -25,12 +25,14 @@ import lombok.ToString;
 public class EpidemicSurveillance {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String name;
     private String idCard;
     private String village;
     private String createTime;
+    // 1, 确诊， 2，治愈，3，隔离，4， 无症状
     @ApiModelProperty(value = "状态", example = "1")
     private Integer state;
+    // 1.男， 2.女
     private Integer sexType;
 }
